added required asterisk
If you´re building a custom block with a form and the form helper it´s easy to set required fields with the validation helper, but it´s not easy to show this requirement to the user. So I´ve given the label function a new param where dev could pass "true" and an asterisk beneath the label is shown.